--- a/src/inflate.ts
+++ b/src/inflate.ts
@@ -14,7 +14,7 @@ import {Uint8WriteStream} from './utils/Uint8WriteStream';
 const FIXED_HUFFMAN_TABLE = generateHuffmanTable( makeFixedHuffmanCodelenValues() );
 
 export function inflate(input: Uint8Array, offset: number = 0) {
-  const buffer = new Uint8WriteStream(BLOCK_MAX_BUFFER_LEN);
+  const buffer = new Uint8WriteStream(input.length * 10);
 
   const stream = new BitReadStream(input, offset);
   let bFinal = 0;