Improve inflate speed.
Test with PNG data 4096 x 4096 (8.2 MB)

### Resuce memory expansion https://github.com/zprodev/zlib.es/commit/b35157dce5c9d2a8fef91cca79a2c0cd14c092d6

|No.|before|after|
|:---:|---:|---:|
|1|28362 ms|1397 ms|
|2|27921 ms|1391 ms|
|3|28546 ms|1363 ms|
|4|28888 ms|1397 ms|
|5|28702 ms|1374 ms|

### Improve bit reading https://github.com/zprodev/zlib.es/commit/7d3e15f6ee5224a9628ea2926a7e177b045933c5

|No.|after|
|:---:|---:|
|1|1267 ms|
|2|1350 ms|
|3|1227 ms|
|4|1259 ms|
|5|1235 ms|